pyproject.toml
[project]
name = "adbc-poolhouse"
version = "0.1.0"
description = "Connection pooling for ADBC drivers from typed warehouse configs"
readme = "README.md"
authors = [
    { name = "Anentropic", email = "ego@anentropic.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "adbc-driver-manager>=1.0.0",
]

[project.optional-dependencies]
duckdb = ["duckdb>=0.9.1"]
snowflake = ["adbc-driver-snowflake>=1.0.0"]
postgresql = ["adbc-driver-postgresql>=1.0.0"]
flightsql = ["adbc-driver-flightsql>=1.0.0"]
bigquery = ["adbc-driver-bigquery>=1.3.0"]
all = [
    "adbc-poolhouse[duckdb]",
    "adbc-poolhouse[snowflake]",
    "adbc-poolhouse[postgresql]",
    "adbc-poolhouse[flightsql]",
    "adbc-poolhouse[bigquery]",
]

[build-system]
requires = ["uv_build>=0.9.18,<1.0.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "basedpyright>=1.38.0",
    "coverage[toml]",
    "ipython>=9.10.0",
    "pdbpp>=0.12.0.post1",
    "pyarrow>=23.0.1",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.1",
    "syrupy>=4.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=0.26.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]

[tool.basedpyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
include = ["src", "tests"]
reportPrivateUsage = false

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH", "D"]
ignore = [
    "D1",    # don't require docstrings everywhere
    "D202",  # allow blank line after docstring (common in tests)
    "D203",  # conflicts with D211 (no blank line before class docstring)
    "D212",  # conflicts with D213 (multiline summary on second line)
    "D401",  # imperative mood is awkward for dunder methods
    "D413",  # blank line after last section not required
]
